[WIP] 0ベクトルを無視する処理を追加

+++ include/render.h
@@ -17,7 +17,7 @@ int draw_vertical_line(
 	void *mlx_addr, // mut
 	t_vec_i32x4 start, // スクリーン上で始点になるベクトル
 	uint32_t vline[],
-	size_t vline_len
+	int vline_len
 );
 
 uint32_t *get_vertical_arr_n(void *img_arr, int index, t_vec_i32x4 size, int out_size);
